requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
idna==2.6
jira==1.0.11
oauthlib==2.0.6
ordereddict==1.1
pbr==3.1.1
python-bugzilla==2.1.0
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
